Cleaned up main.py for unneccessary code, put main function in loop
